requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.25.0
langchain>=0.1.16
langchain-core>=0.1.46
langchain-community>=0.0.32
langchain-ollama>=0.0.6
langgraph>=0.0.30
pinecone>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
pytest>=7.0.0
pre-commit>=3.0.0
black>=23.0.0
flake8>=6.0.0
